+++ src/pages/DuplicantEditorPage/components/DuplicantEditor/components/Attributes/components/AttributeName.tsx
@@ -5,12 +5,14 @@ import { WithTranslation, withTranslation } from "react-i18next";
 import Typography from "@mui/material/Typography";
 
 export interface AttributeNameProps {
+  className?: string;
   attributeId: string;
 }
 
 type Props = AttributeNameProps & WithTranslation;
-const AttributeName: React.FC<Props> = ({ attributeId, t }) => (
+const AttributeName: React.FC<Props> = ({ className, attributeId, t }) => (
   <Typography
+    className={className}
     component="span"
     variant="body1"
     title={t(`oni:DUPLICANTS.ATTRIBUTES.${attributeId}.DESC`, {